+++ catatom2osm.py
@@ -4,6 +4,7 @@ Tool to convert INSPIRE data sets from the Spanish Cadastre ATOM Services to OSM
 """
 import os
 import codecs
+import gzip
 import logging
 import shutil
 from collections import defaultdict, Counter, OrderedDict
@@ -204,7 +205,7 @@ class CatAtom2Osm:
                         report.address_stats(task_osm)
                         report.cons_stats(task_osm)
                         fn = os.path.join('tasks', label + '.osm')
-                        self.write_osm(task_osm, fn)
+                        self.write_osm(task_osm, fn, compress=True)
                         report.osm_stats(task_osm)
                 else:
                     to_clean.append(zone.id())
@@ -346,21 +347,30 @@ class CatAtom2Osm:
                 len(data.relations))
         return data
 
-    def write_osm(self, data, filename):
+    def write_osm(self, data, filename, **kwargs):
         """
         Generates a OSM XML file for a OSM data set.
 
         Args:
             data (Osm): OSM data set
             filename (str): output filename
+            compress (bool): whether the file is to be compressed
+                Defaults to not compressing the file
         """
+        compress = kwargs.pop('compress', False)
         for e in data.elements:
             if 'ref' in e.tags:
                 del e.tags['ref']
-        osm_path = os.path.join(self.path, filename)
         data.merge_duplicated()
-        with codecs.open(osm_path, "w", "utf-8") as file_obj:
-            osmxml.serialize(file_obj, data)
+        if compress:
+            filename += '.gz'
+            osm_path = os.path.join(self.path, filename)
+            file_obj = codecs.EncodedFile(gzip.open(osm_path, "w"), "utf-8")
+        else:
+            osm_path = os.path.join(self.path, filename)
+            file_obj = codecs.open(osm_path, "w", "utf-8")
+        osmxml.serialize(file_obj, data)
+        file_obj.close()
         log.info(_("Generated '%s': %d nodes, %d ways, %d relations"),
             filename, len(data.nodes), len(data.ways), 
             len(data.relations))